Implemented typechecking most things except function falls, typeclasses, and complex types
The design for most of this seems to be holding up fine. There's a bit of a hack to do return type checking which I'm unhappy with, but it's not bad enough to be worth worrying about.

+++ Compiler/TypeChecker/Constraints/Gt.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 
 namespace Speedycloud.Compiler.TypeChecker.Constraints {
     public class Gt : ITypeConstraint{
-        public int Num { get; private set; }
+        public decimal Num { get; private set; }
 
         public override string ToString() {
             return string.Format("(Gt {0})", Num);
         }
 
-        public Gt(int num) {
+        public Gt(decimal num) {
             Num = num;
         }
